backend/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
numpy
pandas
openpyxl
Pillow
phonenumbers
